@@ -1,5 +1,6 @@
 package inaugural.soliloquy.gamestate.test.fakes;
 
+import soliloquy.specs.common.shared.Direction;
 import soliloquy.specs.graphics.assets.ImageAsset;
 import soliloquy.specs.graphics.assets.ImageAssetSet;
 
@@ -25,7 +26,7 @@ public class FakeImageAssetSet implements ImageAssetSet {
     }
 
     @Override
-    public ImageAsset getImageAssetForTypeAndDirection(String s, String s1)
+    public ImageAsset getImageAssetForTypeAndDirection(String s, Direction direction)
             throws IllegalArgumentException {
         return null;
     }
