@@ -2,26 +2,24 @@ package inaugural.soliloquy.gamestate.test.unit.entities;
 
 import inaugural.soliloquy.gamestate.entities.CharacterImpl;
 import inaugural.soliloquy.gamestate.test.fakes.*;
-import inaugural.soliloquy.gamestate.test.stubs.VariableCacheStub;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import soliloquy.specs.common.infrastructure.VariableCache;
 import soliloquy.specs.gamestate.entities.Character;
-import soliloquy.specs.gamestate.entities.*;
 import soliloquy.specs.gamestate.entities.exceptions.EntityDeletedException;
 import soliloquy.specs.gamestate.factories.*;
 import soliloquy.specs.ruleset.entities.CharacterAIType;
-import soliloquy.specs.ruleset.entities.CharacterStaticStatisticType;
 import soliloquy.specs.ruleset.entities.CharacterType;
 
 import java.util.UUID;
 
+import static inaugural.soliloquy.tools.random.Random.randomString;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
+import static soliloquy.specs.common.shared.Direction.SOUTHWEST;
 
 class CharacterImplTests {
-    private Character _character;
-
     private final UUID UUID = java.util.UUID.randomUUID();
     private final CharacterType CHARACTER_TYPE = new FakeCharacterType();
     private final CharacterEventsFactory CHARACTER_EVENTS_FACTORY =
@@ -35,12 +33,16 @@ class CharacterImplTests {
             new FakeCharacterVariableStatisticsFactory();
     private final CharacterStatusEffectsFactory STATUS_EFFECTS_FACTORY =
             new FakeCharacterStatusEffectsFactory();
-    private final VariableCache DATA = new VariableCacheStub();
+
+    @Mock private VariableCache mockData;
+
+    private Character character;
 
     @BeforeEach
     void setUp() {
+        mockData = mock(VariableCache.class);
 
-        _character = new CharacterImpl(
+        character = new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
                 CHARACTER_EVENTS_FACTORY,
@@ -49,7 +51,7 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 ENTITIES_OF_TYPE_FACTORY,
                 STATUS_EFFECTS_FACTORY,
-                DATA);
+                mockData);
     }
 
     @Test
@@ -63,7 +65,7 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 ENTITIES_OF_TYPE_FACTORY,
                 STATUS_EFFECTS_FACTORY,
-                DATA));
+                mockData));
         assertThrows(IllegalArgumentException.class, () -> new CharacterImpl(
                 UUID,
                 null,
@@ -73,7 +75,7 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 ENTITIES_OF_TYPE_FACTORY,
                 STATUS_EFFECTS_FACTORY,
-                DATA));
+                mockData));
         assertThrows(IllegalArgumentException.class, () -> new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
@@ -83,7 +85,7 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 ENTITIES_OF_TYPE_FACTORY,
                 STATUS_EFFECTS_FACTORY,
-                DATA));
+                mockData));
         assertThrows(IllegalArgumentException.class, () -> new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
@@ -93,7 +95,7 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 ENTITIES_OF_TYPE_FACTORY,
                 STATUS_EFFECTS_FACTORY,
-                DATA));
+                mockData));
         assertThrows(IllegalArgumentException.class, () -> new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
@@ -103,7 +105,7 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 ENTITIES_OF_TYPE_FACTORY,
                 STATUS_EFFECTS_FACTORY,
-                DATA));
+                mockData));
         assertThrows(IllegalArgumentException.class, () -> new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
@@ -113,7 +115,7 @@ class CharacterImplTests {
                 null,
                 ENTITIES_OF_TYPE_FACTORY,
                 STATUS_EFFECTS_FACTORY,
-                DATA));
+                mockData));
         assertThrows(IllegalArgumentException.class, () -> new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
@@ -123,7 +125,7 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 null,
                 STATUS_EFFECTS_FACTORY,
-                DATA));
+                mockData));
         assertThrows(IllegalArgumentException.class, () -> new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
@@ -133,7 +135,7 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 ENTITIES_OF_TYPE_FACTORY,
                 null,
-                DATA));
+                mockData));
         assertThrows(IllegalArgumentException.class, () -> new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
@@ -148,12 +150,12 @@ class CharacterImplTests {
 
     @Test
     void testGetInterfaceName() {
-        assertEquals(Character.class.getCanonicalName(), _character.getInterfaceName());
+        assertEquals(Character.class.getCanonicalName(), character.getInterfaceName());
     }
 
     @Test
     void testEquals() {
-        Character character2 = new CharacterImpl(
+        var character2 = new CharacterImpl(
                 UUID,
                 CHARACTER_TYPE,
                 CHARACTER_EVENTS_FACTORY,
@@ -162,182 +164,182 @@ class CharacterImplTests {
                 VARIABLE_STATS_FACTORY,
                 ENTITIES_OF_TYPE_FACTORY,
                 STATUS_EFFECTS_FACTORY,
-                DATA);
-        assertEquals(_character, character2);
+                mockData);
+
+        assertEquals(character, character2);
     }
 
     @Test
     void testCharacterType() {
-        assertSame(CHARACTER_TYPE, _character.type());
+        assertSame(CHARACTER_TYPE, character.type());
     }
 
     @Test
     void testClassifications() {
-        assertNotNull(_character.classifications());
+        assertNotNull(character.classifications());
     }
 
     @Test
     void testPronouns() {
-        assertNotNull(_character.pronouns());
+        assertNotNull(character.pronouns());
     }
 
     @Test
     void testSetAndGetStance() {
-        final String stance = "CharacterStance";
-        _character.setStance(stance);
+        final var stance = randomString();
 
-        assertEquals(stance, _character.getStance());
+        character.setStance(stance);
+
+        assertEquals(stance, character.getStance());
     }
 
     @Test
     void testSetAndGetDirection() {
-        String direction = "CharacterDirection";
-        _character.setDirection(direction);
+        var direction = SOUTHWEST;
 
-        assertEquals(direction, _character.getDirection());
+        character.setDirection(direction);
+
+        assertEquals(direction, character.getDirection());
     }
 
     @Test
     void testSetAndGetSpriteSet() {
-        FakeImageAssetSet imageAssetSet = new FakeImageAssetSet();
-        _character.setImageAssetSet(imageAssetSet);
+        var imageAssetSet = new FakeImageAssetSet();
 
-        assertEquals(imageAssetSet, _character.getImageAssetSet());
+        character.setImageAssetSet(imageAssetSet);
+
+        assertEquals(imageAssetSet, character.getImageAssetSet());
     }
 
     @Test
     void testSetAndGetAIType() {
-        CharacterAIType mockCharacterAIType = mock(CharacterAIType.class);
+        var mockCharacterAIType = mock(CharacterAIType.class);
 
-        assertNull(_character.getAIType());
+        assertNull(character.getAIType());
 
-        _character.setAIType(mockCharacterAIType);
+        character.setAIType(mockCharacterAIType);
 
-        assertSame(mockCharacterAIType, _character.getAIType());
+        assertSame(mockCharacterAIType, character.getAIType());
     }
 
     @Test
     void testSetNullAIType() {
-        assertThrows(IllegalArgumentException.class, () -> _character.setAIType(null));
+        assertThrows(IllegalArgumentException.class, () -> character.setAIType(null));
     }
 
     @Test
     void testCharacterEvents() {
-        assertNotNull(_character.events());
+        assertNotNull(character.events());
     }
 
     @Test
     void testEquipment() {
         assertSame(FakeCharacterEquipmentSlotsFactory.CHARACTER_EQUIPMENT_SLOTS,
-                _character.equipmentSlots());
+                character.equipmentSlots());
     }
 
     @Test
     void testInventory() {
-        assertNotNull(_character.inventory());
+        assertNotNull(character.inventory());
     }
 
     @Test
     void testVariableStatistics() {
-        assertNotNull(_character.variableStatistics());
+        assertNotNull(character.variableStatistics());
     }
 
     @Test
     void testStaticStatistics() {
-        assertNotNull(_character.staticStatistics());
+        assertNotNull(character.staticStatistics());
     }
 
     @Test
     void testStatusEffects() {
-        assertNotNull(_character.statusEffects());
+        assertNotNull(character.statusEffects());
     }
 
     // TODO: Consider testing the ability methods a touch more robustly
     @Test
     void testPassiveAbilities() {
-        assertNotNull(_character.passiveAbilities());
+        assertNotNull(character.passiveAbilities());
     }
 
     @Test
     void testActiveAbilities() {
-        assertNotNull(_character.activeAbilities());
+        assertNotNull(character.activeAbilities());
     }
 
     @Test
     void testReactiveAbilities() {
-        assertNotNull(_character.reactiveAbilities());
+        assertNotNull(character.reactiveAbilities());
     }
 
     @Test
     void testSetAndGetPlayerControlled() {
-        _character.setPlayerControlled(true);
-        assertTrue(_character.getPlayerControlled());
+        character.setPlayerControlled(true);
+        assertTrue(character.getPlayerControlled());
 
-        _character.setPlayerControlled(false);
-        assertFalse(_character.getPlayerControlled());
+        character.setPlayerControlled(false);
+        assertFalse(character.getPlayerControlled());
     }
 
     @Test
     void testData() {
-        assertSame(DATA, _character.data());
+        assertSame(mockData, character.data());
     }
 
     @Test
     void testSetAndGetName() {
-        String CHARACTER_NAME = "CharacterName";
-        _character.setName(CHARACTER_NAME);
+        var name = randomString();
 
-        assertEquals(CHARACTER_NAME, _character.getName());
+        character.setName(name);
+
+        assertEquals(name, character.getName());
     }
 
     @Test
     void testUuid() {
-        assertSame(UUID, _character.uuid());
+        assertSame(UUID, character.uuid());
     }
 
     @Test
     void testAssignCharacterToTile() {
-        Tile tile = new FakeTile();
-        assertNull(_character.tile());
+        var tile = new FakeTile();
+        assertNull(character.tile());
 
         // NB: Character.TILE should NOT be exposed, and calling Character.assignCharacterToTile
         // violates the invariant condition; therefore, FakeTileEntities calls
         // Character.assignCharacterToTile indirectly, as it should be in production code
-        tile.characters().add(_character);
+        tile.characters().add(character);
 
-        assertSame(tile, _character.tile());
+        assertSame(tile, character.tile());
     }
 
     @SuppressWarnings({"rawtypes"})
     @Test
     void testDelete() {
-        Tile tile = new FakeTile();
+        var tile = new FakeTile();
 
-        tile.characters().add(_character);
+        tile.characters().add(character);
 
-        FakeCharacterEquipmentSlots equipmentSlots =
-                (FakeCharacterEquipmentSlots) _character.equipmentSlots();
+        var equipmentSlots = (FakeCharacterEquipmentSlots) character.equipmentSlots();
 
-        FakeCharacterInventory inventory = (FakeCharacterInventory) _character.inventory();
+        var inventory = (FakeCharacterInventory) character.inventory();
 
-        CharacterVariableStatistics variableStats = _character.variableStatistics();
+        var variableStats = character.variableStatistics();
 
-        EntityMembersOfType<CharacterStaticStatisticType,
-                CharacterStatistic<CharacterStaticStatisticType>, Character>
-                staticStats = _character.staticStatistics();
+        var staticStats = character.staticStatistics();
 
-        CharacterStatusEffects statusEffects = _character.statusEffects();
+        var statusEffects = character.statusEffects();
 
-        assertFalse(_character.isDeleted());
-        assertFalse(((FakeTileEntities) tile.characters()).REMOVED_ENTITIES
-                .contains(_character));
+        assertFalse(character.isDeleted());
+        assertFalse(((FakeTileEntities) tile.characters()).REMOVED_ENTITIES.contains(character));
 
-        _character.delete();
+        character.delete();
 
-        assertTrue(_character.isDeleted());
-        Character removedCharacter =
-                (Character) ((FakeTileEntities) tile.characters()).REMOVED_ENTITIES.get(0);
-        assertSame(_character, removedCharacter);
+        assertTrue(character.isDeleted());
+        var removedCharacter = (Character) ((FakeTileEntities) tile.characters()).REMOVED_ENTITIES.get(0);
+        assertSame(character, removedCharacter);
         assertTrue(equipmentSlots._isDeleted);
         assertTrue(inventory._isDeleted);
         assertTrue(variableStats.isDeleted());
@@ -347,78 +349,78 @@ class CharacterImplTests {
 
     @Test
     void testThrowsIllegalStateExceptionWhenDeleted() {
-        _character.delete();
+        character.delete();
 
-        assertThrows(EntityDeletedException.class, () -> _character.type());
-        assertThrows(EntityDeletedException.class, () -> _character.classifications());
-        assertThrows(EntityDeletedException.class, () -> _character.pronouns());
-        assertThrows(EntityDeletedException.class, () -> _character.tile());
-        assertThrows(EntityDeletedException.class, () -> _character.getStance());
-        assertThrows(EntityDeletedException.class, () -> _character.setStance(""));
-        assertThrows(EntityDeletedException.class, () -> _character.getDirection());
-        assertThrows(EntityDeletedException.class, () -> _character.setDirection(""));
-        assertThrows(EntityDeletedException.class, () -> _character.getImageAssetSet());
+        assertThrows(EntityDeletedException.class, () -> character.type());
+        assertThrows(EntityDeletedException.class, () -> character.classifications());
+        assertThrows(EntityDeletedException.class, () -> character.pronouns());
+        assertThrows(EntityDeletedException.class, () -> character.tile());
+        assertThrows(EntityDeletedException.class, () -> character.getStance());
+        assertThrows(EntityDeletedException.class, () -> character.setStance(""));
+        assertThrows(EntityDeletedException.class, () -> character.getDirection());
+        assertThrows(EntityDeletedException.class, () -> character.setDirection(SOUTHWEST));
+        assertThrows(EntityDeletedException.class, () -> character.getImageAssetSet());
         assertThrows(EntityDeletedException.class,
-                () -> _character.setImageAssetSet(new FakeImageAssetSet()));
-        assertThrows(EntityDeletedException.class, () -> _character.getAIType());
-        assertThrows(EntityDeletedException.class, () -> _character.setAIType(null));
-        assertThrows(EntityDeletedException.class, () -> _character.events());
-        assertThrows(EntityDeletedException.class, () -> _character.equipmentSlots());
-        assertThrows(EntityDeletedException.class, () -> _character.inventory());
-        assertThrows(EntityDeletedException.class, () -> _character.variableStatistics());
-        assertThrows(EntityDeletedException.class, () -> _character.staticStatistics());
-        assertThrows(EntityDeletedException.class, () -> _character.statusEffects());
-        assertThrows(EntityDeletedException.class, () -> _character.passiveAbilities());
-        assertThrows(EntityDeletedException.class, () -> _character.activeAbilities());
-        assertThrows(EntityDeletedException.class, () -> _character.reactiveAbilities());
-        assertThrows(EntityDeletedException.class, () -> _character.getPlayerControlled());
-        assertThrows(EntityDeletedException.class, () -> _character.setPlayerControlled(true));
-        assertThrows(EntityDeletedException.class, () -> _character.data());
+                () -> character.setImageAssetSet(new FakeImageAssetSet()));
+        assertThrows(EntityDeletedException.class, () -> character.getAIType());
+        assertThrows(EntityDeletedException.class, () -> character.setAIType(null));
+        assertThrows(EntityDeletedException.class, () -> character.events());
+        assertThrows(EntityDeletedException.class, () -> character.equipmentSlots());
+        assertThrows(EntityDeletedException.class, () -> character.inventory());
+        assertThrows(EntityDeletedException.class, () -> character.variableStatistics());
+        assertThrows(EntityDeletedException.class, () -> character.staticStatistics());
+        assertThrows(EntityDeletedException.class, () -> character.statusEffects());
+        assertThrows(EntityDeletedException.class, () -> character.passiveAbilities());
+        assertThrows(EntityDeletedException.class, () -> character.activeAbilities());
+        assertThrows(EntityDeletedException.class, () -> character.reactiveAbilities());
+        assertThrows(EntityDeletedException.class, () -> character.getPlayerControlled());
+        assertThrows(EntityDeletedException.class, () -> character.setPlayerControlled(true));
+        assertThrows(EntityDeletedException.class, () -> character.data());
         assertThrows(EntityDeletedException.class,
-                () -> _character.assignTileAfterAddedToTileEntitiesOfType(null));
-        assertThrows(EntityDeletedException.class, () -> _character.getName());
-        assertThrows(EntityDeletedException.class, () -> _character.setName(""));
-        assertThrows(EntityDeletedException.class, () -> _character.getInterfaceName());
+                () -> character.assignTileAfterAddedToTileEntitiesOfType(null));
+        assertThrows(EntityDeletedException.class, () -> character.getName());
+        assertThrows(EntityDeletedException.class, () -> character.setName(""));
+        assertThrows(EntityDeletedException.class, () -> character.getInterfaceName());
     }
 
     @SuppressWarnings("rawtypes")
     @Test
     void testEnforceTileInvariant() {
-        FakeTile tile = new FakeTile();
-        tile.characters().add(_character);
-        ((FakeTileEntities) tile.characters()).ENTITIES.remove(_character);
+        var tile = new FakeTile();
+        tile.characters().add(character);
+        ((FakeTileEntities) tile.characters()).ENTITIES.remove(character);
 
-        assertThrows(IllegalStateException.class, () -> _character.type());
-        assertThrows(IllegalStateException.class, () -> _character.classifications());
-        assertThrows(IllegalStateException.class, () -> _character.pronouns());
-        assertThrows(IllegalStateException.class, () -> _character.tile());
-        assertThrows(IllegalStateException.class, () -> _character.getStance());
-        assertThrows(IllegalStateException.class, () -> _character.setStance(""));
-        assertThrows(IllegalStateException.class, () -> _character.getDirection());
-        assertThrows(IllegalStateException.class, () -> _character.setDirection(""));
-        assertThrows(IllegalStateException.class, () -> _character.getImageAssetSet());
+        assertThrows(IllegalStateException.class, () -> character.type());
+        assertThrows(IllegalStateException.class, () -> character.classifications());
+        assertThrows(IllegalStateException.class, () -> character.pronouns());
+        assertThrows(IllegalStateException.class, () -> character.tile());
+        assertThrows(IllegalStateException.class, () -> character.getStance());
+        assertThrows(IllegalStateException.class, () -> character.setStance(""));
+        assertThrows(IllegalStateException.class, () -> character.getDirection());
+        assertThrows(IllegalStateException.class, () -> character.setDirection(SOUTHWEST));
+        assertThrows(IllegalStateException.class, () -> character.getImageAssetSet());
         assertThrows(IllegalStateException.class,
-                () -> _character.setImageAssetSet(new FakeImageAssetSet()));
-        assertThrows(IllegalStateException.class, () -> _character.getAIType());
-        assertThrows(IllegalStateException.class, () -> _character.setAIType(null));
-        assertThrows(IllegalStateException.class, () -> _character.events());
-        assertThrows(IllegalStateException.class, () -> _character.equipmentSlots());
-        assertThrows(IllegalStateException.class, () -> _character.inventory());
-        assertThrows(IllegalStateException.class, () -> _character.variableStatistics());
-        assertThrows(IllegalStateException.class, () -> _character.staticStatistics());
-        assertThrows(IllegalStateException.class, () -> _character.statusEffects());
-        assertThrows(IllegalStateException.class, () -> _character.passiveAbilities());
-        assertThrows(IllegalStateException.class, () -> _character.activeAbilities());
-        assertThrows(IllegalStateException.class, () -> _character.reactiveAbilities());
-        assertThrows(IllegalStateException.class, () -> _character.getPlayerControlled());
-        assertThrows(IllegalStateException.class, () -> _character.setPlayerControlled(true));
-        assertThrows(IllegalStateException.class, () -> _character.data());
-        assertThrows(IllegalStateException.class, () -> _character.delete());
-        assertThrows(IllegalStateException.class, () -> _character.isDeleted());
+                () -> character.setImageAssetSet(new FakeImageAssetSet()));
+        assertThrows(IllegalStateException.class, () -> character.getAIType());
+        assertThrows(IllegalStateException.class, () -> character.setAIType(null));
+        assertThrows(IllegalStateException.class, () -> character.events());
+        assertThrows(IllegalStateException.class, () -> character.equipmentSlots());
+        assertThrows(IllegalStateException.class, () -> character.inventory());
+        assertThrows(IllegalStateException.class, () -> character.variableStatistics());
+        assertThrows(IllegalStateException.class, () -> character.staticStatistics());
+        assertThrows(IllegalStateException.class, () -> character.statusEffects());
+        assertThrows(IllegalStateException.class, () -> character.passiveAbilities());
+        assertThrows(IllegalStateException.class, () -> character.activeAbilities());
+        assertThrows(IllegalStateException.class, () -> character.reactiveAbilities());
+        assertThrows(IllegalStateException.class, () -> character.getPlayerControlled());
+        assertThrows(IllegalStateException.class, () -> character.setPlayerControlled(true));
+        assertThrows(IllegalStateException.class, () -> character.data());
+        assertThrows(IllegalStateException.class, () -> character.delete());
+        assertThrows(IllegalStateException.class, () -> character.isDeleted());
         assertThrows(IllegalStateException.class,
-                () -> _character.assignTileAfterAddedToTileEntitiesOfType(null));
-        assertThrows(IllegalStateException.class, () -> _character.getName());
-        assertThrows(IllegalStateException.class, () -> _character.setName(""));
-        assertThrows(IllegalStateException.class, () -> _character.getInterfaceName());
+                () -> character.assignTileAfterAddedToTileEntitiesOfType(null));
+        assertThrows(IllegalStateException.class, () -> character.getName());
+        assertThrows(IllegalStateException.class, () -> character.setName(""));
+        assertThrows(IllegalStateException.class, () -> character.getInterfaceName());
     }
 }
